About-DataAnalytics/load_docx2json/requirements.txt
numpy==2.2.4
pandas==2.2.3
python-docx==1.1.2
